naturalLanguagePython/CountryService/dictionaryValueInformationFormatter.py
__author__ = 'Antoine'
import re


class DictionaryValueInformationFormatter(object):

    def formatPopulationValue(self, dictionary):
        population = "population"
        if population in dictionary:
            formattedValue = []
            regex = re.compile("\d+(\s\d+){1,}")
            for element in dictionary[population]:
                if regex.match(element):
                    element =element.replace(" ", "")
                formattedValue.append(element)
            dictionary[population] = formattedValue
        return dictionary